remove idhasher (#7)
The behaviour of typeid and hash is not guaranteed. In the future the layout of TypeID may change and the hash impl will either be wrong or begin to panic. And performance of SipHash is fast enough that we won't be impacted on it more than the allocation of our Boxed type and the HashMap anyway